requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
numpy==1.26.4
openpyxl==3.1.2
python-dotenv==1.0.1 